Add files via upload
I added lists and stylised the main heading with blue colour as well as adding additional headings. I also changed the topic of the website to reflect on how Storror, a parkour group have greatly inspired me (idk if that's correct but u can always change the content as needs be).